+++ scripts/run_collect_and_compile.sh
@@ -8,44 +8,8 @@
 
 set -e
 
-export ORA_BENCH_MULTIPLE_RUN=true
-
-if [ -z "$ORA_BENCH_RUN_CX_ORACLE_PYTHON" ]; then
-    export ORA_BENCH_RUN_CX_ORACLE_PYTHON=true
-fi
-if [ -z "$ORA_BENCH_RUN_GODROR_GO" ]; then
-    export ORA_BENCH_RUN_GODROR_GO=true
-fi
-if [ -z "$ORA_BENCH_RUN_JAMDB_ORACLE_ERLANG" ]; then
-    export ORA_BENCH_RUN_JAMDB_ORACLE_ERLANG=true
-fi
-if [ -z "$ORA_BENCH_RUN_JDBC_JAVA" ]; then
-    export ORA_BENCH_RUN_JDBC_JAVA=true
-fi
-if [ -z "$ORA_BENCH_RUN_ODPI_C" ]; then
-    export ORA_BENCH_RUN_ODPI_C=true
-fi
-if [ -z "$ORA_BENCH_RUN_ORANIF_ELIXIR" ]; then
-    export ORA_BENCH_RUN_ORANIF_ELIXIR=true
-fi
-if [ -z "$ORA_BENCH_RUN_ORANIF_ERLANG" ]; then
-    export ORA_BENCH_RUN_ORANIF_ERLANG=true
-fi
-
-if [ -z "$ORA_BENCH_FILE_CONFIGURATION_NAME" ]; then
-    export ORA_BENCH_FILE_CONFIGURATION_NAME=priv/properties/ora_bench.properties
-fi
-
 if [ -z "$GOPATH" ]; then
     GOPATH=$(pwd)/src_go/go
-fi
-
-if [ -z "$ORA_BENCH_JAVA_CLASSPATH" ]; then
-    if [ "$OSTYPE" = "msys" ]; then
-        export ORA_BENCH_JAVA_CLASSPATH=".;priv/java_jar/*"
-    else
-        export ORA_BENCH_JAVA_CLASSPATH=".:priv/java_jar/*"
-    fi
 fi
 
 if [ -z "$RUN_GLOBAL_JAMDB" ]; then
@@ -60,26 +24,24 @@ echo "Start $0"
 echo "--------------------------------------------------------------------------------"
 echo "ora_bench - Oracle benchmark - collect libraries and compile."
 echo "--------------------------------------------------------------------------------"
-echo "BULKFILE_EXISTING          : $BULKFILE_EXISTING"
+echo "BULKFILE_EXISTING                 : $BULKFILE_EXISTING"
 echo "--------------------------------------------------------------------------------"
-echo "RUN_GLOBAL_JAMDB           : $RUN_GLOBAL_JAMDB"
-echo "RUN_GLOBAL_NON_JAMDB       : $RUN_GLOBAL_NON_JAMDB"
+echo "RUN_GLOBAL_JAMDB                  : $RUN_GLOBAL_JAMDB"
+echo "RUN_GLOBAL_NON_JAMDB              : $RUN_GLOBAL_NON_JAMDB"
 echo "--------------------------------------------------------------------------------"
-echo "RUN_CX_ORACLE_PYTHON       : $ORA_BENCH_RUN_CX_ORACLE_PYTHON"
-echo "RUN_GODROR_GO              : $ORA_BENCH_RUN_GODROR_GO"
-echo "RUN_JAMDB_ORACLE_ERLANG    : $ORA_BENCH_RUN_JAMDB_ORACLE_ERLANG"
-echo "RUN_JDBC_JAVA              : $ORA_BENCH_RUN_JDBC_JAVA"
-echo "RUN_ODPI_C                 : $ORA_BENCH_RUN_ODPI_C"
-echo "RUN_ORANIF_ELIXIR          : $ORA_BENCH_RUN_ORANIF_ELIXIR"
-echo "RUN_ORANIF_ERLANG          : $ORA_BENCH_RUN_ORANIF_ERLANG"
+echo "RUN_CX_ORACLE_PYTHON              : $ORA_BENCH_RUN_CX_ORACLE_PYTHON"
+echo "RUN_JDBC_KOTLIN                   : $ORA_BENCH_RUN_JDBC_KOTLIN"
+echo "RUN_GODROR_GO                     : $ORA_BENCH_RUN_GODROR_GO"
+echo "RUN_JAMDB_ORACLE_ERLANG           : $ORA_BENCH_RUN_JAMDB_ORACLE_ERLANG"
+echo "RUN_JDBC_JAVA                     : $ORA_BENCH_RUN_JDBC_JAVA"
+echo "RUN_ODPI_C                        : $ORA_BENCH_RUN_ODPI_C"
+echo "RUN_ORANIF_ELIXIR                 : $ORA_BENCH_RUN_ORANIF_ELIXIR"
+echo "RUN_ORANIF_ERLANG                 : $ORA_BENCH_RUN_ORANIF_ERLANG"
 echo "--------------------------------------------------------------------------------"
-echo "FILE_CONFIGURATION_NAME    : $ORA_BENCH_FILE_CONFIGURATION_NAME"
-echo "--------------------------------------------------------------------------------"
-echo "GOPATH                     : $GOPATH"
-echo "GOROOT                     : $GOROOT"
-echo "GRADLE_HOME                : $GRADLE_HOME"
-echo "JAVA_CLASSPATH             : $ORA_BENCH_JAVA_CLASSPATH"
-echo "LD_LIBRARY_PATH            : $LD_LIBRARY_PATH"
+echo "GOPATH                            : $GOPATH"
+echo "GOROOT                            : $GOROOT"
+echo "GRADLE_HOME                       : $GRADLE_HOME"
+echo "LD_LIBRARY_PATH                   : $LD_LIBRARY_PATH"
 echo "--------------------------------------------------------------------------------"
 date +"DATE TIME : %d.%m.%Y %H:%M:%S"
 echo "================================================================================"
@@ -93,7 +55,7 @@ fi
 if [ "$RUN_GLOBAL_NON_JAMDB" = "true" ]; then
     if [ "$ORA_BENCH_RUN_ODPI_C" == "true" ]; then
         echo "Setup C - Start ============================================================" 
-        if ! java -cp "$ORA_BENCH_JAVA_CLASSPATH" ch.konnexions.orabench.OraBench setup_c; then
+        if ! java -jar priv/libs/ora_bench_java.jar setup_c; then
             exit 255
         fi
 
@@ -117,7 +79,7 @@ if [ "$RUN_GLOBAL_NON_JAMDB" = "true" ]; then
     
     if [ "$ORA_BENCH_RUN_ORANIF_ELIXIR" == "true" ]; then
         echo "Setup Elixir - Start =======================================================" 
-        if ! java -cp "$ORA_BENCH_JAVA_CLASSPATH" ch.konnexions.orabench.OraBench setup_elixir; then
+        if ! java -jar priv/libs/ora_bench_java.jar setup_elixir; then
             exit 255
         fi
 
@@ -153,7 +115,7 @@ fi
     
 if [ "$ORA_BENCH_RUN_JAMDB_ORACLE_ERLANG" == "true" ] || [ "$ORA_BENCH_RUN_ORANIF_ERLANG" == "true" ]; then
     echo "Setup Erlang - Start ======================================================="
-    if ! java -cp "$ORA_BENCH_JAVA_CLASSPATH" ch.konnexions.orabench.OraBench setup_erlang; then
+    if ! java -jar priv/libs/ora_bench_java.jar setup_erlang; then
         exit 255
     fi
     
@@ -180,9 +142,17 @@ if [ "$RUN_GLOBAL_NON_JAMDB" = "true" ]; then
         echo "Setup Go - End   ===========================================================" 
     fi    
 
+    if [ "$ORA_BENCH_RUN_JDBC_KOTLIN" == "true" ]; then
+        echo "Setup Kotlin - Start =======================================================" 
+        if ! ./src_kotlin/scripts/run_gradle.sh; then
+            exit 255
+        fi
+        echo "Setup Kotlin - End   =======================================================" 
+    fi    
+
     if [ "$ORA_BENCH_RUN_CX_ORACLE_PYTHON" == "true" ]; then
         echo "Setup Python - Start =======================================================" 
-        if ! java -cp "priv/java_jar/*" ch.konnexions.orabench.OraBench setup_python; then
+        if ! java -jar priv/libs/ora_bench_java.jar setup_python; then
             exit 255
         fi
         echo "Setup Python - End   ======================================================="