+++ scripts/run_create_bulk_file.sh
@@ -12,24 +12,14 @@ if [ -z "$ORA_BENCH_FILE_CONFIGURATION_NAME" ]; then
     export ORA_BENCH_FILE_CONFIGURATION_NAME=priv/properties/ora_bench.properties
 fi
 
-if [ -z "$ORA_BENCH_JAVA_CLASSPATH" ]; then
-    if [ "$OSTYPE" = "msys" ]; then
-        export ORA_BENCH_JAVA_CLASSPATH=".;priv/java_jar/*"
-    else
-        export ORA_BENCH_JAVA_CLASSPATH=".:priv/java_jar/*"
-    fi
-fi
-
 echo "================================================================================"
 echo "Start $0"
 echo "--------------------------------------------------------------------------------"
 echo "ora_bench - Oracle benchmark - setup benchmark run."
 echo "--------------------------------------------------------------------------------"
-echo "MULTIPLE_RUN               : $ORA_BENCH_MULTIPLE_RUN"
+echo "CHOICE_DRIVER                     : $ORA_BENCH_CHOICE_DRIVER"
 echo "--------------------------------------------------------------------------------"
-echo "FILE_CONFIGURATION_NAME    : $ORA_BENCH_FILE_CONFIGURATION_NAME"
-echo "--------------------------------------------------------------------------------"
-echo "JAVA_CLASSPATH             : $ORA_BENCH_JAVA_CLASSPATH"
+echo "FILE_CONFIGURATION_NAME           : $ORA_BENCH_FILE_CONFIGURATION_NAME"
 echo "--------------------------------------------------------------------------------"
 date +"DATE TIME : %d.%m.%Y %H:%M:%S"
 echo "================================================================================"
@@ -44,7 +34,7 @@ if ! { /bin/bash src_java/scripts/run_gradle.sh; }; then
     exit 255
 fi
 
-if ! java -cp "$ORA_BENCH_JAVA_CLASSPATH" ch.konnexions.orabench.OraBench setup; then
+if ! java -jar priv/libs/ora_bench_java.jar setup; then
     exit 255
 fi