+++ scripts/run_finalise_benchmark.sh
@@ -12,29 +12,15 @@ if [ -z "$ORA_BENCH_FILE_CONFIGURATION_NAME" ]; then
     export ORA_BENCH_FILE_CONFIGURATION_NAME=priv/properties/ora_bench.properties
 fi
 
-if [ -z "$ORA_BENCH_JAVA_CLASSPATH" ]; then
-    if [ "$OSTYPE" = "msys" ]; then
-        export ORA_BENCH_JAVA_CLASSPATH=".;priv/java_jar/*"
-    else
-        export ORA_BENCH_JAVA_CLASSPATH=".:priv/java_jar/*"
-    fi
-fi
-
 echo "================================================================================"
 echo "Start $0"
 echo "--------------------------------------------------------------------------------"
 echo "ora_bench - Oracle benchmark - finalise benchmark run."
 echo "--------------------------------------------------------------------------------"
-echo "MULTIPLE_RUN               : $ORA_BENCH_MULTIPLE_RUN"
-echo "--------------------------------------------------------------------------------"
-echo "FILE_CONFIGURATION_NAME    : $ORA_BENCH_FILE_CONFIGURATION_NAME"
-echo "--------------------------------------------------------------------------------"
-echo "JAVA_CLASSPATH             : $ORA_BENCH_JAVA_CLASSPATH"
-echo "--------------------------------------------------------------------------------"
 date +"DATE TIME : %d.%m.%Y %H:%M:%S"
 echo "================================================================================"
 
-if ! java -cp "priv/java_jar/*" ch.konnexions.orabench.OraBench finalise; then
+if ! java -jar priv/libs/ora_bench_java.jar finalise; then
     exit 255
 fi