+++ src_elixir/scripts/run_bench_oranif.sh
@@ -10,7 +10,7 @@ if [ -z "$ORA_BENCH_BENCHMARK_DATABASE" ]; then
     export ORA_BENCH_BENCHMARK_DATABASE=db_19_3_ee
 fi
 if [ -z "$ORA_BENCH_CONNECTION_HOST" ]; then
-    export ORA_BENCH_CONNECTION_HOST=0.0.0.0
+    export javaORA_BENCH_CONNECTION_HOST=localhost
 fi
 if [ -z "$ORA_BENCH_CONNECTION_PORT" ]; then
     export ORA_BENCH_CONNECTION_PORT=1521
@@ -21,14 +21,6 @@ fi
 
 if [ -z "$ORA_BENCH_FILE_CONFIGURATION_NAME" ]; then
     export ORA_BENCH_FILE_CONFIGURATION_NAME=priv/properties/ora_bench.properties
-fi
-
-if [ -z "$ORA_BENCH_JAVA_CLASSPATH" ]; then
-    if [ "$OSTYPE" = "msys" ]; then
-        export ORA_BENCH_JAVA_CLASSPATH=".;priv/java_jar/*"
-    else
-        export ORA_BENCH_JAVA_CLASSPATH=".:priv/java_jar/*"
-    fi
 fi
 
 echo "================================================================================"
@@ -49,8 +41,6 @@ echo "BENCHMARK_TRANSACTION_SIZE : $ORA_BENCH_BENCHMARK_TRANSACTION_SIZE"
 echo "--------------------------------------------------------------------------------"
 echo "FILE_CONFIGURATION_NAME    : $ORA_BENCH_FILE_CONFIGURATION_NAME"
 echo "--------------------------------------------------------------------------------"
-echo "JAVA_CLASSPATH             : $ORA_BENCH_JAVA_CLASSPATH"
-echo "--------------------------------------------------------------------------------"
 date +"DATE TIME : %d.%m.%Y %H:%M:%S"
 echo "================================================================================"
 
@@ -59,7 +49,7 @@ if [ "$ORA_BENCH_MULTIPLE_RUN" != "true" ]; then
         exit 255
     fi
 
-    if ! java -cp "priv/java_jar/*" ch.konnexions.orabench.OraBench setup_elixir; then
+    if ! java -jar priv/libs/ora_bench_java.jar setup_elixir; then
         exit 255
     fi
 fi    
@@ -67,14 +57,15 @@ fi
 (
     cd src_elixir || exit 255
     
-    if [ -f "mix.lock" ]; then
-        rm -f mix.lock
-    fi         
-    if [ -f "deps" ]; then
-        rm -rf deps
-    fi         
-
     if [ "$ORA_BENCH_MULTIPLE_RUN" != "true" ]; then
+
+        if [ -f "mix.lock" ]; then
+            rm -f mix.lock
+        fi
+        if [ -f "deps" ]; then
+            rm -rf deps
+        fi
+
         if ! mix local.hex --force; then
             exit 255
         fi
@@ -82,9 +73,11 @@ fi
         if ! mix deps.clean --all; then
             exit 255
         fi
+
         if ! mix deps.get; then
             exit 255
         fi
+
         if ! mix deps.compile; then
             exit 255
         fi