+++ src_kotlin/src/main/kotlin/ch/konnexions/Insert.kt
@@ -1,0 +1,69 @@
+package ch.konnexions
+
+import java.sql.Connection
+import java.sql.PreparedStatement
+
+import org.apache.log4j.Logger
+
+class Insert(
+    logger: Logger,
+    isDebug: Boolean,
+    connection: Connection,
+    preparedStatement: PreparedStatement,
+    bulkDataPartition: ArrayList<Array<String>>,
+    benchmarkSize: Int,
+    benchmarkTransactionSize: Int
+) :
+    Runnable {
+    private val benchmarkSize: Int
+    private val benchmarkTransactionSize: Int
+    private val bulkDataPartition: ArrayList<Array<String>>
+    private val connection: Connection
+    private val isDebug: Boolean
+    private val logger: Logger
+    private val preparedStatement: PreparedStatement
+
+    /**
+     * Runs the thread implementer.
+     */
+    override fun run() {
+        if (isDebug) {
+            logger.debug("Start")
+        }
+
+        OraBench.insertHelper(
+            logger,
+            isDebug,
+            connection,
+            preparedStatement,
+            bulkDataPartition,
+            benchmarkSize,
+            benchmarkTransactionSize
+        )
+
+        if (isDebug) {
+            logger.debug("End")
+        }
+    }
+
+    /**
+     * Instantiates a new Insert class.
+     */
+    init {
+        if (isDebug) {
+            logger.debug("Start")
+        }
+
+        this.benchmarkSize = benchmarkSize
+        this.benchmarkTransactionSize = benchmarkTransactionSize
+        this.bulkDataPartition = bulkDataPartition
+        this.connection = connection
+        this.isDebug = isDebug
+        this.logger = logger
+        this.preparedStatement = preparedStatement
+
+        if (isDebug) {
+            logger.debug("End")
+        }
+    }
+}